pyproject.toml
[project]
name = "pycrypto"
version = "0.1.0"
description = ""
authors = [{ name="Wesley Silva", email="wesleysilv@gmail.com"}]
readme = "README.md"
requires-python = ">=3.13"
package-mode = false
dependencies = [
    "dotenv (>=0.9.9,<0.10.0)",
    "numpy (>=2.2,<3.0)",
    "binance-connector (>=3.12.0,<4.0.0)",
    "flask (>=3.1.2,<4.0.0)",
    "psycopg[binary] (>=3.3.2,<4.0.0)",
    "redis (>=7.1.0,<8.0.0)",
    "logging (>=0.4.9.6,<0.5.0.0)",
    "ta-lib (>=0.6.8,<0.7.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"


[dependency-groups]
dev = [
    "pytest (>=9.0.2,<10.0.0)",
    "pytest-watch (>=4.2.0,<5.0.0)",
    "pytest-cov (>=7.0.0,<8.0.0)",
    "pytest-dotenv (>=0.5.2,<0.6.0)",
    "pylint (>=4.0.4,<5.0.0)",
    "icecream (>=2.1.8,<3.0.0)",
    "taskipy (>=1.14.1,<2.0.0)"
]
doc = [
    "mkdocs (>=1.6.1,<2.0.0)",
    "mkdocstrings (>=1.0.0,<2.0.0)",
    "mkdocstrings-python (>=2.0.1,<3.0.0)",
    "mkdocs-material (>=9.7.1,<10.0.0)"
]


[tool.poetry.dependencies]
python = ">=3.13,<4.0"


[tool.pytest.ini_options]
pythonpath = "."
testpaths = "tests"
python_files = "test_*.py"
env_files = ".env"
addopts = "--doctest-modules"
markers = ["delete_db_data: Tests who can delete some info on postgres database."]


[tool.isort]
profile = "black"
line_length = 79


[tool.taskipy.tasks]
lint = "isort --check --diff . && pylint app/"
pre_test = "task lint"
test = "pytest tests/* -s -m 'not delete_db_data' --cov=app -vv"
post_test = "coverage html"
docs = "mkdocs serve"

